requirements.txt
wget
farm-haystack[milvus,preprocessing]
fastapi
uvicorn[standard]